Rename VASP example file for Al_fcc in debyetools
This commit introduces a new example file, DOSCAR.EvV.10, for the FCC aluminum structure in the debyetools package. This addition can be used to validate and test calculations related to material properties within the package. The format of the data adheres to the standard CAR file specifications used in density of states applications. This example helps expand the library of sample cases available for users of the debyetools, enhancing its utility for academic and practical applications.