Update version and fix energy calculation bug
Updated the package version from 2.7.3 to 2.7.4 in `setup.py` to reflect the changes in functionality. Fixed a bug in `load_data_from_DFT.py` where the energy calculation did not correctly pass the `current_folder` parameter to the `get_energy` function within the loop. This ensures the correct handling of potential files. The version bump signifies this bug fix and maintains versioning consistency. These changes improve the reliability and accuracy of the overall thermodynamic property calculations.